test_backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.14
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.4.0
django-filter==24.3
django-js-asset==2.2.0
django-mptt==0.14.0
django-mptt-admin==2.6.2
django-sass-processor==1.4
django-treebeard==4.7.1
djangorestframework==3.15.2
djangorestframework-recursive==0.1.2
libsass==0.23.0
pillow==10.3.0
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.3
typing_extensions==4.12.2
tzdata==2022.4
